MOVEONLY: CBlockUndo from main.h to undo.h
